+++ CHANGELOG.md
@@ -14,6 +14,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   - Binary search:
     - No longer compares a non-existent value at index array.length
     - Compares against the matcher object more consistently
+- Collection
+  - Predicate can now be deeply nested.
 
 ## [0.3.1] - 2019-06-20
 - Fixed Tests